fix(runtime-error): pull in canary packages for react and styles (#33)
### Related Ticket(s)

Unhandled Runtime Error #32

### Description

NextJS error overlay appears when there is an uncaught error. The uncaught error was coming from our Profile service, so we caught the error and console logged it out instead. Updating to use the canary version os the react and styles packages pulls in the latest changes in the services.

### Changelog

**Changed**

- update to canary versions of ibmdotcom react and styles packages