fix(env): add sass path variable to build script (#46)
### Description

This change sets the `SASS_PATH` environment variable from the `build` script. This will prevent any sass-related build errors when `.env` is missing or, if present, does not contain an entry for the path. This does not interfere with applications already using `.env` with the entry present.

Tested in both Mac/Win environments.

### Changelog

**New**

- `cross-env` package
- add windows-specific syntax for `SASS_PATH` to README

**Changed**

- add `cross-env` to `build` script in package.json
- re-organized `package.json`